Move helper method into hook
